Add spinner decorator and update error message in functions
A spinner decorator was added to show progress CLI data fills while it's executing. An error message within functions.py was also updated to reference a constant value instead of hard coded string. This enhances the user experience as they can visually see that a process is ongoing, while also improving the maintainability of the code.